+++ src/quantizer.cc
@@ -22,12 +22,12 @@ void Dequantize(std::array<int16_t, 16>& coefficients, int16_t DCfact,
 
 using namespace internal;
 
-static const size_t kClampLb = 0;
-static const size_t kClampUb = 127;
+static const int kClampLb = 0;
+static const int kClampUb = 127;
 
 void QuantizeY(std::array<int16_t, 16>& coefficients, int16_t qp,
                const QuantIndices& quantizer_header) {
-  size_t idx = std::clamp(size_t(qp + quantizer_header.y_dc_delta_q), kClampLb,
+  size_t idx = std::clamp(qp + quantizer_header.y_dc_delta_q, kClampLb,
                           kClampUb);
   int16_t DCfact = std::clamp(int16_t(kDClookup.at(idx)), int16_t(8), int16_t(132));
   int16_t ACfact =
@@ -37,9 +37,9 @@ void QuantizeY(std::array<int16_t, 16>& coefficients, int16_t qp,
 
 void QuantizeUV(std::array<int16_t, 16>& coefficients, int16_t qp,
                 const QuantIndices& quantizer_header) {
-  size_t idx_dc = std::clamp(size_t(qp + quantizer_header.uv_dc_delta_q),
+  size_t idx_dc = std::clamp(qp + quantizer_header.uv_dc_delta_q,
                              kClampLb, kClampUb);
-  size_t idx_ac = std::clamp(size_t(qp + quantizer_header.uv_ac_delta_q),
+  size_t idx_ac = std::clamp(qp + quantizer_header.uv_ac_delta_q,
                              kClampLb, kClampUb);
   int16_t DCfact = std::clamp(int16_t(kDClookup.at(idx_dc)), int16_t(8), int16_t(132));
   int16_t ACfact = std::clamp(int16_t(kAClookup.at(idx_ac)), int16_t(8), int16_t(132));
@@ -48,9 +48,9 @@ void QuantizeUV(std::array<int16_t, 16>& coefficients, int16_t qp,
 
 void QuantizeY2(std::array<int16_t, 16>& coefficients, int16_t qp,
                 const QuantIndices& quantizer_header) {
-  size_t idx_dc = std::clamp(size_t(quantizer_header.y2_dc_delta_q + qp),
+  size_t idx_dc = std::clamp(quantizer_header.y2_dc_delta_q + qp,
                              kClampLb, kClampUb);
-  size_t idx_ac = std::clamp(size_t(quantizer_header.y2_ac_delta_q + qp),
+  size_t idx_ac = std::clamp(quantizer_header.y2_ac_delta_q + qp,
                              kClampLb, kClampUb);
   int16_t DCfact =
       std::clamp(int16_t(kDClookup.at(idx_dc) * 2), int16_t(8), int16_t(132));
@@ -61,7 +61,7 @@ void QuantizeY2(std::array<int16_t, 16>& coefficients, int16_t qp,
 
 void DequantizeY(std::array<int16_t, 16>& coefficients, int16_t qp,
                  const QuantIndices& quantizer_header) {
-  size_t idx = std::clamp(size_t(quantizer_header.y_dc_delta_q + qp), kClampLb,
+  size_t idx = std::clamp(quantizer_header.y_dc_delta_q + qp, kClampLb,
                           kClampUb);
   int16_t DCfact = int16_t(kDClookup.at(idx));
   int16_t ACfact = int16_t(kAClookup.at(size_t(qp)));
@@ -70,9 +70,9 @@ void DequantizeY(std::array<int16_t, 16>& coefficients, int16_t qp,
 
 void DequantizeUV(std::array<int16_t, 16>& coefficients, int16_t qp,
                   const QuantIndices& quantizer_header) {
-  size_t idx_dc = std::clamp(size_t(quantizer_header.uv_dc_delta_q + qp),
+  size_t idx_dc = std::clamp(quantizer_header.uv_dc_delta_q + qp,
                              kClampLb, kClampUb);
-  size_t idx_ac = std::clamp(size_t(quantizer_header.uv_ac_delta_q + qp),
+  size_t idx_ac = std::clamp(quantizer_header.uv_ac_delta_q + qp,
                              kClampLb, kClampUb);
   int16_t DCfact = int16_t(kDClookup.at(idx_dc));
   if (DCfact > 132) DCfact = 132;
@@ -82,13 +82,17 @@ void DequantizeUV(std::array<int16_t, 16>& coefficients, int16_t qp,
 
 void DequantizeY2(std::array<int16_t, 16>& coefficients, int16_t qp,
                   const QuantIndices& quantizer_header) {
-  size_t idx_dc = std::clamp(size_t(quantizer_header.y2_dc_delta_q + qp),
+  size_t idx_dc = std::clamp(quantizer_header.y2_dc_delta_q + qp,
                              kClampLb, kClampUb);
-  size_t idx_ac = std::clamp(size_t(quantizer_header.y2_ac_delta_q + qp),
+  size_t idx_ac = std::clamp(quantizer_header.y2_ac_delta_q + qp,
                              kClampLb, kClampUb);
   int16_t DCfact = int16_t(kDClookup.at(idx_dc) * 2);
   int16_t ACfact = int16_t(kAClookup.at(idx_ac) * 155 / 100);
   if (ACfact < 8) ACfact = 8;
+  // if (qp == 0) {
+    // std::cerr << "AC = " << ACfact << std::endl;
+    // std::cerr << "Delta = " << int(quantizer_header.y2_ac_delta_q) << std::endl;
+  // }
   Dequantize(coefficients, DCfact, ACfact);
 }
